+++ compiler/src/org.graalvm.compiler.nodes/src/org/graalvm/compiler/nodes/java/TypeSwitchNode.java
@@ -40,6 +40,7 @@ import org.graalvm.compiler.nodes.AbstractBeginNode;
 import org.graalvm.compiler.nodes.ConstantNode;
 import org.graalvm.compiler.nodes.FixedWithNextNode;
 import org.graalvm.compiler.nodes.NodeView;
+import org.graalvm.compiler.nodes.ProfileData.SwitchProbabilityData;
 import org.graalvm.compiler.nodes.ValueNode;
 import org.graalvm.compiler.nodes.extended.LoadHubNode;
 import org.graalvm.compiler.nodes.extended.SwitchNode;
@@ -62,10 +63,11 @@ public final class TypeSwitchNode extends SwitchNode implements LIRLowerable, Si
     protected final ResolvedJavaType[] keys;
     protected final Constant[] hubs;
 
-    public TypeSwitchNode(ValueNode value, AbstractBeginNode[] successors, ResolvedJavaType[] keys, double[] keyProbabilities, int[] keySuccessors, ConstantReflectionProvider constantReflection) {
-        super(TYPE, value, successors, keySuccessors, keyProbabilities);
+    public TypeSwitchNode(ValueNode value, AbstractBeginNode[] successors, ResolvedJavaType[] keys, int[] keySuccessors, ConstantReflectionProvider constantReflection,
+                    SwitchProbabilityData profileData) {
+        super(TYPE, value, successors, keySuccessors, profileData);
         assert successors.length <= keys.length + 1;
-        assert keySuccessors.length == keyProbabilities.length;
+        assert keySuccessors.length == profileData.getKeyProbabilities().length;
         this.keys = keys;
         assert value.stamp(NodeView.DEFAULT) instanceof AbstractPointerStamp;
         assert assertKeys();
@@ -196,7 +198,9 @@ public final class TypeSwitchNode extends SwitchNode implements LIRLowerable, Si
                     }
 
                     AbstractBeginNode[] successorsArray = newSuccessors.toArray(new AbstractBeginNode[newSuccessors.size()]);
-                    TypeSwitchNode newSwitch = graph().add(new TypeSwitchNode(value(), successorsArray, newKeys, newKeyProbabilities, newKeySuccessors, tool.getConstantReflection()));
+                    TypeSwitchNode newSwitch = graph().add(
+                                    new TypeSwitchNode(value(), successorsArray, newKeys, newKeySuccessors, tool.getConstantReflection(),
+                                                    SwitchProbabilityData.create(newKeyProbabilities, profileData.getProfileSource())));
                     ((FixedWithNextNode) predecessor()).setNext(newSwitch);
                     GraphUtil.killWithUnusedFloatingInputs(this);