+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/heap/ReferenceInternals.java
@@ -24,7 +24,7 @@
  */
 package com.oracle.svm.core.heap;
 
-import static org.graalvm.compiler.nodes.extended.BranchProbabilityNode.LUDICROUSLY_FAST_PATH_PROBABILITY;
+import static org.graalvm.compiler.nodes.extended.BranchProbabilityNode.EXTREMELY_FAST_PATH_PROBABILITY;
 import static org.graalvm.compiler.nodes.extended.BranchProbabilityNode.probability;
 
 import java.lang.ref.Reference;
@@ -236,7 +236,7 @@ public final class ReferenceInternals {
 
     public static void updateSoftReferenceClock() {
         long now = TimeUtils.divideNanosToMillis(System.nanoTime()); // should be monotonous, ensure
-        if (probability(LUDICROUSLY_FAST_PATH_PROBABILITY, now >= Target_java_lang_ref_SoftReference.clock)) {
+        if (probability(EXTREMELY_FAST_PATH_PROBABILITY, now >= Target_java_lang_ref_SoftReference.clock)) {
             Target_java_lang_ref_SoftReference.clock = now;
         }
     }