+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/fieldfolding/StaticFinalFieldFoldingFeature.java
@@ -95,11 +95,11 @@ import jdk.vm.ci.meta.ResolvedJavaField;
  * a static final field as written and does not perform constant folding. Without constant folding
  * during parsing already, other graph builder plugins like
  * {@link IntrinsifyMethodHandlesInvocationPlugin} do not work on such fields.
- * 
+ *
  * This feature performs constant folding for a limited but important class of static final fields:
  * the class initializer contains a single field store and the stored value is a constant. That
  * single constant is propagated to field loads.
- * 
+ *
  * The specification of Java class initializers and static final fields complicate this
  * optimization: Even if it is guaranteed that the field eventually gets a constant value assigned
  * in the class initializer, field loads that happen before the field store while the class
@@ -110,7 +110,7 @@ import jdk.vm.ci.meta.ResolvedJavaField;
  * slow-path case of returning the uninitialized value. All these boolean values are stored in the
  * {@link #fieldInitializationStatus} array, and {@link #fieldCheckIndexMap} stores the index for
  * the optimized fields.
- * 
+ *
  * The optimized field load is also preceded by a {@link EnsureClassInitializedNode} to trigger the
  * necessary class initialization. It would be possible to combine the class initialization check
  * and the field initialization check to a single check. But that leads to several corner cases and
@@ -382,7 +382,7 @@ final class StaticFinalFieldFoldingNodePlugin implements NodePlugin {
 
         EndNode uninitializedEndNode = b.getGraph().add(new EndNode());
         EndNode initializedEndNode = b.getGraph().add(new EndNode());
-        b.add(new IfNode(isUninitializedNode, uninitializedEndNode, initializedEndNode, BranchProbabilityNode.LUDICROUSLY_SLOW_PATH_PROBABILITY));
+        b.add(new IfNode(isUninitializedNode, uninitializedEndNode, initializedEndNode, BranchProbabilityNode.EXTREMELY_SLOW_PATH_PROFILE));
 
         MergeNode merge = b.append(new MergeNode());
         merge.addForwardEnd(uninitializedEndNode);